refactor: add libs directory
